fix: Renamed detectable results to keep consistent convention of result names describing what went wrong as supposed to what to do next (IDETECT-2196).
