--- detectable/src/main/java/com/synopsys/integration/detectable/detectable/result/NpmRunInstallDetectableResult.java
@@ -1,36 +1,0 @@
-/**
- * detectable
- *
- * Copyright (c) 2020 Synopsys, Inc.
- *
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership. The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package com.synopsys.integration.detectable.detectable.result;
-
-public class NpmRunInstallDetectableResult extends FailedDetectableResult {
-    private final String directoryPath;
-
-    public NpmRunInstallDetectableResult(final String directoryPath) {
-        this.directoryPath = directoryPath;
-    }
-
-    @Override
-    public String toDescription() {
-        return String.format("A package.json was located in %s, but the node_modules folder was NOT located. Please run 'npm install' in that location and try again.", directoryPath);
-    }
-}
