+++ detectable/src/main/java/com/synopsys/integration/detectable/detectable/result/PoetryLockfileNotFoundDetectableResult.java
@@ -22,15 +22,15 @@
  */
 package com.synopsys.integration.detectable.detectable.result;
 
-public class CargoGenerateLockfileDetectResult extends FailedDetectableResult {
+public class PoetryLockfileNotFoundDetectableResult extends FailedDetectableResult {
     private final String directoryPath;
 
-    public CargoGenerateLockfileDetectResult(final String directoryPath) {
+    public PoetryLockfileNotFoundDetectableResult(final String directoryPath) {
         this.directoryPath = directoryPath;
     }
 
     @Override
     public String toDescription() {
-        return String.format("A Cargo.toml was located in %s, but the Cargo.lock file was NOT located. Please run 'cargo generate-lockfile' in that location and try again.", directoryPath);
+        return String.format("A pyproject.toml was located in %s, but the Poetry.lock file was NOT located. Please run 'poetry install' in that location and try again.", directoryPath);
     }
 }
